requirements.txt
contourpy==1.1.1
cycler==0.12.1
filelock==3.15.3
fonttools==4.53.0
fsspec==2024.6.0
importlib-resources==6.4.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.0.3
patsy==0.5.6
pillow==10.3.0
plotly==5.22.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
statsmodels==0.14.1
sympy==1.12.1
tenacity==8.4.1
threadpoolctl==3.5.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
triton==2.3.1
typing-extensions==4.12.2
tzdata==2024.1
zipp==3.19.2
